Add caching scripts
We were kicking the sh*t out of the OAC arduino by having three or four
clients open at once just querying for statuses. As a result, we now only ask
for a status up date once every five minutes, or when the web interface locks
or unlocks. This is Nice™